requirements/base.txt
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
rcssmin==1.1.1  # https://github.com/ndparker/rcssmin
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
redis==4.5.4  # https://github.com/redis/redis-py
hiredis==2.2.2  # https://github.com/redis/hiredis-py
celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
uvicorn[standard]==0.22.0  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==4.1.9  # pyup: < 4.2  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.54.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-compressor==4.3.1  # https://github.com/django-compressor/django-compressor
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.0.0  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.26.2  # https://github.com/tfranzel/drf-spectacular

# my installed requirements
django-mongodb-engine==0.6.0
django-quill-editor==0.1.40
djongo==1.3.6
